moved get-material
