+++ src/scene/structure/parametric-objects/common/get-material.js
