+++ test/scene/structure/architecture-toolkit/window.js
@@ -1,4 +1,4 @@
-import applyDefaults from '../../../../src/scene/structure/apply-defaults.js';
+import applyDefaults from '../../../../src/scene/structure/apply-defaults.js'
 import window from '../../../../src/scene/structure/parametric-objects/window'
 
 // mock runtime module to prevent from tests blowing up
@@ -9,7 +9,7 @@ test('get window data3d', async function(){
   const data3d = await window(attributes)
 
   // mesh names
-  expect(Object.keys(data3d)).toEqual([ 'frame', 'glass' ])
+  expect(Object.keys(data3d.meshes)).toEqual([ 'frame', 'glass' ])
   expect(data3d.meshes.frame.positions).toBeDefined()
   expect(data3d.meshes.frame.normals).toBeDefined()
 });
